capstone_environment/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.51.0
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.11
django-bootstrap4==22.2
django-bootstrap5==22.1
django-crispy-forms==1.14.0
django-forms-bootstrap==3.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-form-bootstrap-4==0.5.2
et-xmlfile==1.1.0
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.23.2
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.4
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stem==1.8.0
tomli==2.0.1
torrequest==0.1.0
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.12
